requirements.txt
gpiozero~=1.5.1
Pillow~=8.0.1
numpy~=1.19.4
scipy~=1.5.4
sounddevice~=0.4.1
paho_mqtt~=1.5.1